fix(chunk): onboarding chunk path
