+++ views/profiles.pug
@@ -28,3 +28,4 @@ html
       };
     script(src=webpackConfig.output.publicPath + '/test/integration/bulle-extension.js')
     script(src=webpackConfig.output.publicPath + '/js/profiles.bundle.js')
+    script(src=webpackConfig.output.publicPath + '/js/OnBoarding.chunk.js')
